requirements.txt
streamlit
PyPDF2
sentence-transformers
google-generativeai
speechrecognition
textstat
python-dotenv